refactor: migrate estudiantes data to content collections (#265)
Resolves #256
