+++ src/content/estudiantes_accesos/aula-virtual.json
@@ -1,0 +1,8 @@
+{
+  "id": "aula-virtual",
+  "nombre": "Aula Virtual",
+  "descripcion": "Plataforma de aprendizaje en línea",
+  "icono": "monitor",
+  "href": "https://aulavirtual.unapiquitos.edu.pe",
+  "color": "bg-epg-navy"
+}
