+++ src/content/estudiantes_accesos/sigae.json
@@ -1,0 +1,8 @@
+{
+  "id": "sigae",
+  "nombre": "SIGAE",
+  "descripcion": "Sistema Integrado de Gestión Académica",
+  "icono": "clipboard",
+  "href": "https://sigae.unapiquitos.edu.pe",
+  "color": "bg-epg-navy"
+}
