+++ src/content/estudiantes_calendario/inicio-clases-2026-1.json
@@ -1,0 +1,8 @@
+{
+  "id": "inicio-clases-2026-1",
+  "titulo": "Inicio de Clases 2026-I",
+  "descripcion": "Inicio del semestre académico 2026-I",
+  "fechaInicio": "2026-06-01",
+  "tipo": "evento",
+  "importante": true
+}
