+++ src/content/estudiantes_documentos/formato-proyecto-tesis.json
@@ -1,0 +1,9 @@
+{
+  "id": "formato-proyecto-tesis",
+  "nombre": "Formato de Proyecto de Tesis",
+  "descripcion": "Plantilla oficial para la presentación del proyecto de investigación",
+  "tipo": "docx",
+  "categoria": "formato",
+  "url": "/documentos/formato-proyecto-tesis.docx",
+  "fechaActualizacion": "2024-01-10"
+}
