+++ src/content/estudiantes_documentos/guia-elaboracion-tesis.json
@@ -1,0 +1,9 @@
+{
+  "id": "guia-elaboracion-tesis",
+  "nombre": "Guía de Elaboración de Tesis",
+  "descripcion": "Manual con lineamientos para la redacción de tesis",
+  "tipo": "pdf",
+  "categoria": "guia",
+  "url": "/documentos/guia-elaboracion-tesis.pdf",
+  "fechaActualizacion": "2024-03-01"
+}
