+++ src/content/estudiantes_documentos/manual-apa7.json
@@ -1,0 +1,9 @@
+{
+  "id": "manual-apa7",
+  "nombre": "Manual de Estilo APA 7ma Edición",
+  "descripcion": "Guía de citación y referencias bibliográficas",
+  "tipo": "pdf",
+  "categoria": "manual",
+  "url": "/documentos/manual-apa7.pdf",
+  "fechaActualizacion": "2023-12-01"
+}
