+++ src/content/estudiantes_documentos/reglamento-general.json
@@ -1,0 +1,9 @@
+{
+  "id": "reglamento-general",
+  "nombre": "Reglamento General de la EPG",
+  "descripcion": "Normativa general que rige el funcionamiento de la Escuela de Postgrado",
+  "tipo": "pdf",
+  "categoria": "reglamento",
+  "url": "/documentos/reglamento-general-epg.pdf",
+  "fechaActualizacion": "2024-03-15"
+}
