+++ src/content/estudiantes_documentos/reglamento-grados.json
@@ -1,0 +1,9 @@
+{
+  "id": "reglamento-grados",
+  "nombre": "Reglamento de Grados y Títulos",
+  "descripcion": "Procedimientos para la obtención de grados académicos",
+  "tipo": "pdf",
+  "categoria": "reglamento",
+  "url": "/documentos/reglamento-grados-titulos.pdf",
+  "fechaActualizacion": "2024-02-20"
+}
