+++ src/content/estudiantes_faq/faq-1.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-1",
+  "pregunta": "¿Cuál es el plazo máximo para terminar la maestría?",
+  "respuesta": "El plazo máximo para culminar los estudios de maestría es de 4 años contados desde la fecha de ingreso. En casos excepcionales, se puede solicitar una ampliación de plazo presentando una solicitud debidamente justificada.",
+  "categoria": "general"
+}
