+++ src/content/estudiantes_faq/faq-2.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-2",
+  "pregunta": "¿Cómo puedo solicitar una constancia de estudios?",
+  "respuesta": "Debe presentar una solicitud (FUT) en mesa de partes adjuntando el recibo de pago correspondiente (S/ 30.00). El documento estará listo en 3 a 5 días hábiles.",
+  "categoria": "tramites"
+}
