+++ src/content/estudiantes_faq/faq-3.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-3",
+  "pregunta": "¿Cuántas veces puedo sustentar mi tesis si desapruebo?",
+  "respuesta": "El estudiante tiene derecho a dos (2) oportunidades adicionales para sustentar su tesis en caso de desaprobación. Cada nueva sustentación requiere un pago adicional por derecho de sustentación.",
+  "categoria": "tesis"
+}
