+++ src/content/estudiantes_faq/faq-4.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-4",
+  "pregunta": "¿Puedo realizar pagos fraccionados?",
+  "respuesta": "Sí, la pensión de estudios puede pagarse en cuotas mensuales según el cronograma establecido por la EPG. El incumplimiento de pagos puede generar intereses y restricciones en el acceso a servicios académicos.",
+  "categoria": "pagos"
+}
