+++ src/content/estudiantes_faq/faq-5.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-5",
+  "pregunta": "¿Qué porcentaje de similitud acepta Turnitin?",
+  "respuesta": "El porcentaje máximo de similitud permitido es del 25%. Si la tesis supera este porcentaje, deberá realizar las correcciones necesarias antes de poder sustentar.",
+  "categoria": "tesis"
+}
