+++ src/content/estudiantes_faq/faq-6.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-6",
+  "pregunta": "¿Cómo solicito cambio de asesor de tesis?",
+  "respuesta": "Debe presentar una solicitud (FUT) indicando los motivos del cambio, adjuntando una carta de aceptación del nuevo asesor propuesto. La solicitud será evaluada por el Comité de Investigación.",
+  "categoria": "tesis"
+}
