+++ src/content/estudiantes_faq/faq-7.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-7",
+  "pregunta": "¿Puedo matricularme en asignaturas de otro programa?",
+  "respuesta": "Sí, puede llevar asignaturas electivas de otros programas siempre que existan vacantes disponibles y cuente con la autorización del coordinador de su programa.",
+  "categoria": "matricula"
+}
