+++ src/content/estudiantes_faq/faq-8.json
@@ -1,0 +1,6 @@
+{
+  "id": "faq-8",
+  "pregunta": "¿Qué pasa si no me matriculo en un semestre?",
+  "respuesta": "Si no se matricula en un semestre sin haber solicitado licencia, será considerado como abandono de estudios. Para reingresar, deberá solicitar reincorporación y regularizar su situación académica.",
+  "categoria": "matricula"
+}
